Additonal furnaces, code quality stuff

+++ src/main/java/com/mcmoddev/mineralogy/blocks/Rock.java
@@ -16,8 +16,8 @@ public class Rock extends net.minecraft.block.Block {
 			SoundType sound) {
 		super(Material.ROCK);
 		this.isStoneEquivalent = isStoneEquivalent;
-		this.setHardness((float) hardness); // dirt is 0.5, grass is 0.6, stone is 1.5,iron ore is 3, obsidian is 50
-		this.setResistance((float) blastResistance); // dirt is 0, iron ore is 5, stone is 10, obsidian is 2000
+		this.setHardness(hardness); // dirt is 0.5, grass is 0.6, stone is 1.5,iron ore is 3, obsidian is 50
+		this.setResistance(blastResistance); // dirt is 0, iron ore is 5, stone is 10, obsidian is 2000
 		this.setSoundType(sound); // sound for stone
 		this.setHarvestLevel("pickaxe", toolHardnessLevel);
 	}
